% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorio.R
\name{area_municipios}
\alias{area_municipios}
\title{Returns a \code{data.frame} with area by municipality (in \code{km^2})}
\usage{
area_municipios(dir = ".")
}
\arguments{
\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Returns a \code{data.frame} with area by municipality (in \code{km^2})
}
\examples{
\dontrun{
  df <- area_municipios("~/tmp")
}
}
\seealso{
\url{http://www.ibge.gov.br/home/geociencias/cartografia/default_territ_area.shtm}
}

