% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcb.R
\name{bcb_serie_carrega}
\alias{bcb_serie_carrega}
\title{Returns a \code{data.frame} with a given series from BCB
(Brazil's Central Bank - Banco Central do Brasil) by its code}
\usage{
bcb_serie_carrega(codigo, dir = ".", alwaysDownload = F)
}
\arguments{
\item{codigo}{the series code}

\item{dir}{directory for temporary (cache) files}

\item{alwaysDownload}{if true, ignores local cache}
}
\value{
a \code{data.frame}
}
\description{
Returns a \code{data.frame} with a given series from BCB
(Brazil's Central Bank - Banco Central do Brasil) by its code
}
\examples{
\dontrun{
  df <- bcb_serie_carrega(193)
}
}
\seealso{
\url{https://www.bcb.gov.br/?SGS}
}
