% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea.R
\name{ipea_serie_carrega}
\alias{ipea_serie_carrega}
\title{Returns a \code{data.frame} with a given series from IPEA
(Instituto de Pesquisa Econômica Aplicada) by its code}
\usage{
ipea_serie_carrega(codigo, dir = ".", alwaysDownload = F)
}
\arguments{
\item{codigo}{the series code}

\item{dir}{directory for temporary (cache) files}

\item{alwaysDownload}{if true, ignores local cache}
}
\value{
a \code{data.frame}
}
\description{
Returns a \code{data.frame} with a given series from IPEA
(Instituto de Pesquisa Econômica Aplicada) by its code
}
\examples{
\dontrun{
  df <- ipea_serie_carrega(37667)
}
}
\seealso{
\url{http://www.ipeadata.gov.br}
}
