% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib.R
\name{pib_municipios}
\alias{pib_municipios}
\title{Returns a \code{data.frame} with the gross domestic product (GDP/PIB) by municipality}
\usage{
pib_municipios(ano = NA, dir = ".")
}
\arguments{
\item{ano}{year of requested data (if not supplied, most current information will be returned)}

\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Returns a \code{data.frame} with the gross domestic product (GDP/PIB) by municipality
}
\examples{
\dontrun{
  df <- pib_municipios(2013)
}
}

