% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populacao_estimativa.R
\name{populacao_municipios}
\alias{populacao_municipios}
\title{Returns a \code{data.frame} with population estimates for all municipalities for the requested year.}
\usage{
populacao_municipios(ano, dir = ".")
}
\arguments{
\item{ano}{year of the requested data file}

\item{dir}{directory for temporary files}
}
\value{
a \code{data.frame}
}
\description{
This function internally uses both \code{populacao_municipios_download} & \code{populacao_municipios_carrega}
functions.
}
\examples{
\dontrun{
  df <- populacao_municipios(2014)
}
}

