% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populacao_estimativa.R
\name{populacao_municipios_carrega}
\alias{populacao_municipios_carrega}
\title{Reads a file path with the population estimate file and returns a \code{data.frame}}
\usage{
populacao_municipios_carrega(filename, dir = ".", skip = 2)
}
\arguments{
\item{filename}{the file path}

\item{dir}{directory for temporary files}

\item{skip}{number of header lines to skip}
}
\value{
a \code{data.frame}
}
\description{
Reads a file path with the population estimate file and returns a \code{data.frame}
}
\examples{
\dontrun{
  df <- populacao_municipios_carrega(filename)
}
}
