% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populacao_estimativa.R
\name{populacao_municipios_download}
\alias{populacao_municipios_download}
\title{Returns a file path of the population estimate file for a given year}
\usage{
populacao_municipios_download(ano, dir = ".")
}
\arguments{
\item{ano}{year (2000+) of the requested data file}

\item{dir}{directory for temporary files}
}
\value{
a file path
}
\description{
Returns a file path of the population estimate file for a given year
}
\examples{
\dontrun{
  filename <- populacao_municipios_download(2012)
}
}
