% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflacao.R
\name{precos_deflatorpib}
\alias{precos_deflatorpib}
\title{Returns a \code{data.frame} with yearly GDP deflator}
\usage{
precos_deflatorpib(dir = ".")
}
\arguments{
\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Years avaiable: 1948+
}
\examples{
\dontrun{
  df <- precos_deflatorpib()
}
}
