% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflacao.R
\name{precos_ipca}
\alias{precos_ipca}
\title{Returns a \code{data.frame} with montly IPCA index for the requested year}
\usage{
precos_ipca(ano, dir = NULL)
}
\arguments{
\item{ano}{year (1991+) of requested data}

\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Returns a \code{data.frame} with montly IPCA index for the requested year
}
\examples{
\dontrun{
  df <- precos_ipca(2015)
}
}
