% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflacao.R
\name{precos_sinapi}
\alias{precos_sinapi}
\title{Returns a \code{data.frame} with montly SINAPI index for the requested year}
\usage{
precos_sinapi(ano, dir = NULL)
}
\arguments{
\item{ano}{year (1986+) of requested data}

\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Returns a \code{data.frame} with montly SINAPI index for the requested year
}
\examples{
\dontrun{
  df <- precos_sinapi(2015)
}
}
