% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ribge}
\alias{ribge}
\alias{ribge-package}
\title{ribge: package for (down)loading data from IBGE.}
\description{
ribge is a package for (down)loading data from IBGE - Instituto Brasileiro de Geografia e Estatística.
}
\details{
This package has no affiliation to IBGE.
}
\seealso{
\url{http://www.ibge.gov.br/} & \url{https://github.com/tbrugz/ribge}
}

