% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{series_estatisticas_carrega}
\alias{series_estatisticas_carrega}
\title{Returns a \code{data.frame} with a given series.}
\usage{
series_estatisticas_carrega(codigo, dir = ".", xtraurl = "",
  transpose = FALSE)
}
\arguments{
\item{codigo}{the series code}

\item{dir}{directory for temporary (cache) files}

\item{xtraurl}{extra parameter to include in the request URL}

\item{transpose}{transpose the \code{data.frame}}
}
\value{
a \code{data.frame}
}
\description{
Returns a \code{data.frame} with a given series.
}
\examples{
\dontrun{
  df <- series_estatisticas_carrega("CD99_BR_ABS", transpose=T)
}
}
\seealso{
\url{http://seriesestatisticas.ibge.gov.br/}
}
