% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{series_estatisticas_carrega_todas_localidades}
\alias{series_estatisticas_carrega_todas_localidades}
\title{Returns a \code{data.frame} with a given series for all localilies. Useful when series' geographic level
(nivel) is distinct from 'BR'.}
\usage{
series_estatisticas_carrega_todas_localidades(
  codigo,
  dir = ".",
  xtraurl = "",
  transpose = FALSE
)
}
\arguments{
\item{codigo}{the series code}

\item{dir}{directory for temporary (cache) files}

\item{xtraurl}{extra parameter to include in the request URL}

\item{transpose}{transpose the \code{data.frame}}
}
\value{
a \code{data.frame}
}
\description{
Returns a \code{data.frame} with a given series for all localilies. Useful when series' geographic level
(nivel) is distinct from 'BR'.
}
\examples{
\dontrun{
  df <- series_estatisticas_carrega_todas_localidades("PE62_RM_PERC", transpose=T)
}
}
\seealso{
\url{http://seriesestatisticas.ibge.gov.br/}
}
