% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriesEstatisticas}
\alias{seriesEstatisticas}
\title{Series codes from \code{seriesestatisticas} website}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 702 rows and 5 columns.}
\usage{
seriesEstatisticas
}
\description{
Series codes from \code{seriesestatisticas} website
}
\seealso{
\url{http://seriesestatisticas.ibge.gov.br/}
}
\keyword{datasets}

