% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse.R
\name{tse_municipios}
\alias{tse_municipios}
\title{Returns a \code{data.frame} with municipalities from TSE (Tribunal Superior Eleitoral)}
\usage{
tse_municipios(dir = ".")
}
\arguments{
\item{dir}{directory for temporary files}
}
\value{
the \code{data.frame}
}
\description{
Returns a \code{data.frame} with municipalities from TSE (Tribunal Superior Eleitoral)
}
\examples{
\dontrun{
  df <- tse_municipios()
}
}

