% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tseMunicipios}
\alias{tseMunicipios}
\title{Municipalities from TSE website}
\description{
Municipalities from TSE website
}
\seealso{
\url{https://dadosabertos.tse.jus.br/dataset/eleitorado-atual}
}
