% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.download}
\alias{util.download}
\title{Download a given file}
\usage{
util.download(url, dir = NULL, file = NULL, alwaysDownload = F)
}
\arguments{
\item{url}{file's url to download}

\item{dir}{directory for temporary (cache) files}

\item{file}{name for the local file}

\item{alwaysDownload}{if true, ignores local cache}
}
\value{
a file path
}
\description{
Download a given file
}

