% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.downloadSeries}
\alias{util.downloadSeries}
\title{Downloads a given series by its code}
\usage{
util.downloadSeries(codigo, dir = ".", xtraurl = "")
}
\arguments{
\item{codigo}{the series code}

\item{dir}{directory for temporary (cache) files}

\item{xtraurl}{extra parameter to include in the request URL}
}
\value{
a file path
}
\description{
Downloads a given series by its code
}
\examples{
\dontrun{
  filename <- util.downloadSeries("CD99_BR_ABS")
}
}
\seealso{
\url{http://seriesestatisticas.ibge.gov.br/}
}

