% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.unzip}
\alias{util.unzip}
\title{Unzips a file}
\usage{
util.unzip(src, exdir, unzip = getOption("unzip"))
}
\arguments{
\item{src}{the file to unzip}

\item{exdir}{directory to extract file}

\item{unzip}{unzip strategy}
}
\description{
Unzips a file
}
\seealso{
\url{https://github.com/hadley/devtools/commit/1b1732cc2305a1880e3a788a1160fe85de37b06e}
}
