% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vswitch}
\alias{vswitch}
\title{Vectorized \code{switch}}
\usage{
vswitch(expr, ...)
}
\arguments{
\item{expr}{the expression to evaluate}

\item{...}{extra parameters for \code{switch}}
}
\description{
Vectorized \code{switch}
}
\seealso{
\code{\link{switch}} & \url{http://stackoverflow.com/a/25989828/616413}
}

