% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibge.populacao.sources}
\alias{ibge.populacao.sources}
\title{Municipalities population estimates file sources}
\description{
A \code{data.frame} with municipalities population file sources by year. Internally
used by \code{\link{populacao_municipios_download}} & \code{\link{populacao_municipios}}
}
