% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriesEstatisticas}
\alias{seriesEstatisticas}
\title{Series codes from \code{seriesestatisticas} website}
\description{
Series codes from \code{seriesestatisticas} website
}
\seealso{
\url{http://seriesestatisticas.ibge.gov.br/}
}
